backend/conlang_tools/pyproject.toml
[tool.poetry]
name = "conlang-tools"
version = "0.1.0"
description = ""
authors = ["Joseph Bochinski <stirgejr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
case-converter = "^1.1.0"
sqlalchemy = "^2.0.36"
pyyaml = "^6.0.2"
ptpython = "^3.0.29"
rich = "^13.9.4"
prompt-toolkit = "^3.0.48"
argcomplete = "^3.5.2"
nanoid = "^2.0.0"
gunicorn = "^23.0.0"
flask = "^3.1.0"
flask-cors = "^5.0.0"
cw-logging = "^0.1.10"
nltk = "^3.9.1"
numpy = "^2.2.1"
requests = "^2.32.3"
phonecodes = "^1.1.4"
replbase = "^0.0.28"
gevent = "^25.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
conlang = "conlang_tools.lt_cli.lt_cli:main"
conlang-server = "conlang_tools.server:main"